requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-multipart==0.0.9

pypdf==4.2.0
python-docx==1.1.2
tiktoken==0.7.0

chromadb==0.4.24

openai==1.51.0
google-generativeai==0.7.2
sentence-transformers==3.0.1
numpy==1.26.4

tenacity==8.3.0
requests==2.32.3
httpx==0.27.2
pytest==8.3.2